+++ src/main/java/de/jeff_media/LumberJack/CommandLumberjack.java
@@ -4,17 +4,20 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Objects;
 
 public class CommandLumberjack implements CommandExecutor {
 	
-	LumberJack plugin;
+	final LumberJack plugin;
 	
-	public CommandLumberjack(LumberJack plugin) {
+	CommandLumberjack(LumberJack plugin) {
 		this.plugin=plugin;
 	}
 
 	@Override
-	public boolean onCommand(CommandSender sender, Command command, String alias, String[] args) {
+	public boolean onCommand(@NotNull CommandSender sender, Command command, @NotNull String alias, String[] args) {
 		
 		
 		
@@ -27,7 +30,7 @@ public class CommandLumberjack implements CommandExecutor {
 		}
 		Player p = (Player) sender;
 		if(!sender.hasPermission("lumberjack.use")) {
-			sender.sendMessage(plugin.getCommand("lumberjack").getPermissionMessage());
+			sender.sendMessage(Objects.requireNonNull(Objects.requireNonNull(plugin.getCommand("lumberjack")).getPermissionMessage()));
 			return true;
 		}