+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 - Added support for blocks with upper-diagonal blocks like acacia or dark oak
 - Added support for crimson and warped fungus trees
 - Removed tree-types config option. Let me know if you need this again
+- Added automatic config updater
+- Improved plugin update checker
 - Added Spanish and Turkish translation
+- Updated Spigot API to 1.16.1
+- Code cleanup